regqregerg

+++ web/app/plugins/sttv/includes/functions/sttv-webhook-functions.php
@@ -39,9 +39,11 @@ function trial_expiration_checker() {
 
     if ( !empty( $garbage ) ) {
         foreach ( $garbage as $g ) {
-            $umeta = get_user_meta( $g['wp_id'], 'sttv_user_data', true );
-            $customer = \Stripe\Customer::retrieve( $umeta['customer'] );
-            $customer->delete();
+            if ( $g['wp_id'] > 1 ) {
+                $umeta = get_user_meta( $g['wp_id'], 'sttv_user_data', true );
+                $customer = \Stripe\Customer::retrieve( $umeta['customer'] );
+                $customer->delete();
+            }
             $wpdb->delete( $wpdb->prefix.'trial_reference',
                 [
                     'invoice_id' => $g['invoice_id']
